pyproject.toml
[project]
name = "ai-research-digest"
version = "0.1.0"
description = "Daily AI research paper + tweet digest"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.1",
    "anthropic>=0.40",
    "tweepy>=4.14",
    "notion-client>=2.2",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dateutil>=2.9",
]
